fix: address PR review comments — error messages, logging, and test coverage
- Add subject and object_type as structured fields to MyGrants error log
- Use %v instead of %s for error formatting in performReadTuples
- Return sanitized client-safe errors from MakeServiceUnavailable calls
- Extract performReadTuples unit tests into access_service_test.go and
  remove MyGrants tests that covered logic now in performReadTuples
- Delete my_grants_test.go, merging all service tests into one file
